Bug fix in abi.decode and update of execution tests in related directories (#254)
* revert the previous abi.decode modification

* Updated failing execution tests: semanticTests/{abiEncoderV1, abiEncoderV2, abiencodedecode}.

* Found some out-of-scope tests.

* Updated failing compilation tests.

+++ test/libsolidity/semanticTests/abiEncoderV1/return_dynamic_types_cross_call_advanced.sol
@@ -1,7 +1,7 @@
 contract C {
 	function dyn() public returns (bytes memory a, uint b, bytes20[] memory c, uint d) {
 		a = "1234567890123456789012345678901234567890";
-		b = type(uint).max;
+		b = type(uint256).max;
 		c = new bytes20[](4);
 		c[0] = bytes20(uint160(1234));
 		c[3] = bytes20(uint160(6789));
@@ -15,4 +15,4 @@ contract C {
 // compileViaYul: also
 // EVMVersion: >homestead
 // ----
-// f() -> 0x80, -1, 0xe0, 0x1234, 40, "12345678901234567890123456789012", "34567890", 4, 97767552542602192590433234714624, 0, 0, 537879995309340587922569878831104
+// f() -> "1234567890123456789012345678901234567890", 0x00ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff, refargs { 0x01, 0x04, 0x00000000000000000000000000000000000004d2, 0x0000000000000000000000000000000000000000, 0x0000000000000000000000000000000000000000, 0x1a85 } , 0x1234
